+++ src/tme/models/group_layer.rs
@@ -7,7 +7,7 @@ use super::property::Property;
 #[derive(Debug, Serialize, Deserialize, PartialEq)]
 #[serde(rename_all = "lowercase")]
 pub struct GroupLayer {
-    pub id:         i64,
+    pub id:         i32,
     pub layers:     Vec<Layer>,
     pub name:       String,
     #[serde(rename = "offsetx")]
@@ -17,12 +17,12 @@ pub struct GroupLayer {
     pub opacity:    f64,
     pub properties: Option<Vec<Property>>,
     #[serde(rename = "startx")]
-    pub start_x:    Option<i64>,
+    pub start_x:    Option<i32>,
     #[serde(rename = "starty")]
-    pub start_y:    Option<i64>,
+    pub start_y:    Option<i32>,
     pub visible:    bool,
-    pub x:          i64,
-    pub y:          i64,
+    pub x:          i32,
+    pub y:          i32,
 }
 
 #[cfg(test)]
